Separacion de Interfaces en archivos propios
