+++ Mhotivo/App_Data/Repositories/AcademicYearRepository.cs
@@ -2,22 +2,11 @@
 using System.Data.Entity;
 using System.Linq;
 using System.Linq.Expressions;
+using Mhotivo.App_Data.Repositories.Interfaces;
 using Mhotivo.Models;
 
 namespace Mhotivo.App_Data.Repositories
 {
-    public interface IAcademicYearRepository : IDisposable
-    {
-        AcademicYear First(Expression<Func<AcademicYear, AcademicYear>> query);
-        AcademicYear GetById(long id);
-        AcademicYear Create(AcademicYear itemToCreate);
-        IQueryable<AcademicYear> Query(Expression<Func<AcademicYear, AcademicYear>> expression);
-        IQueryable<AcademicYear> Filter(Expression<Func<AcademicYear, bool>> expression);
-        AcademicYear Update(AcademicYear itemToUpdate, bool updateCourse, bool updateGrade, bool updateTeacher);
-        AcademicYear Delete(long id);
-        void SaveChanges();
-    }
-
     public class AcademicYearRepository : IAcademicYearRepository
     {
         private readonly MhotivoContext _context;
