+++ Mhotivo/App_Data/Repositories/ContactRepository.cs
@@ -4,22 +4,11 @@ using System.Data.Entity;
 using System.Linq;
 using System.ComponentModel;
 using System.Linq.Expressions;
+using Mhotivo.App_Data.Repositories.Interfaces;
 using Mhotivo.Models;
 
 namespace Mhotivo.App_Data.Repositories
 {
-    public interface IContactInformationRepository : IDisposable
-    {
-        ContactInformation First(Expression<Func<ContactInformation, ContactInformation>> query);
-        ContactInformation GetById(long id);
-        ContactInformation Create(ContactInformation itemToCreate);
-        IQueryable<TResult> Query<TResult>(Expression<Func<ContactInformation, TResult>> expression);
-        IQueryable<ContactInformation> Filter(Expression<Func<ContactInformation, bool>> expression);
-        ContactInformation Update(ContactInformation itemToUpdate);
-        ContactInformation Delete(long Id);
-        void SaveChanges();
-    }
-
     public class ContactInformationRepository : IContactInformationRepository
     {
         private readonly MhotivoContext _context;
