+++ Mhotivo/App_Data/Repositories/CourseRepository.cs
@@ -3,22 +3,11 @@ using System.Data.Entity;
 using System.Linq;
 using System.ComponentModel;
 using System.Linq.Expressions;
+using Mhotivo.App_Data.Repositories.Interfaces;
 using Mhotivo.Models;
 
 namespace Mhotivo.App_Data.Repositories
 {
-    public interface ICourseRepository :  IDisposable
-    {
-        Course First(Expression<Func<Course, Course>> query);
-        Course GetById(long id);
-        Course Create(Course itemToCreate);
-        IQueryable<TResult> Query<TResult>(Expression<Func<Course, TResult>> expression);
-        IQueryable<Course> Filter(Expression<Func<Course, bool>> expression);
-        Course Update(Course itemToUpdate);
-        void Delete(Course itemToDelete);
-        void SaveChanges();
-    }
-
     public class CourseRepository : ICourseRepository
     {
         private readonly MhotivoContext _context;
