+++ Mhotivo/App_Data/Repositories/EnrollRepository.cs
@@ -2,22 +2,11 @@
 using System.Data.Entity;
 using System.Linq;
 using System.Linq.Expressions;
+using Mhotivo.App_Data.Repositories.Interfaces;
 using Mhotivo.Models;
 
 namespace Mhotivo.App_Data.Repositories
 {
-    public interface IEnrollRepository
-    {
-        Enroll First(Expression<Func<Enroll, Enroll>> query);
-        Enroll GetById(long id);
-        Enroll Create(Enroll itemToCreate);
-        IQueryable<Enroll> Query(Expression<Func<Enroll, Enroll>> expression);
-        IQueryable<Enroll> Filter(Expression<Func<Enroll, bool>> expression);
-        Enroll Update(Enroll itemToUpdate, bool academicYear, bool student);
-        Enroll Delete(long id);
-        void SaveChanges();
-    }
-
     public class EnrollRepository : IEnrollRepository
     {
         private readonly MhotivoContext _context;
