+++ Mhotivo/App_Data/Repositories/GradeRepository.cs
@@ -3,22 +3,11 @@ using System.Data.Entity;
 using System.Data.Entity.Migrations;
 using System.Linq;
 using System.Linq.Expressions;
+using Mhotivo.App_Data.Repositories.Interfaces;
 using Mhotivo.Models;
 
 namespace Mhotivo.App_Data.Repositories
 {
-    public interface IGradeRepository
-    {
-        Grade First(Expression<Func<Grade, bool>> query);
-        Grade GetById(long id);
-        Grade Create(Grade itemToCreate);
-        IQueryable<Grade> Query(Expression<Func<Grade, Grade>> expression);
-        IQueryable<Grade> Filter(Expression<Func<Grade, bool>> expression);
-        Grade Update(Grade itemToUpdate);
-        void Delete(Grade itemToDelete);
-        void SaveChanges();
-    }
-
     public class GradeRepository : IGradeRepository
     {
         private readonly MhotivoContext _context;
