+++ Mhotivo/App_Data/Repositories/Interfaces/IEnrollRepository.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Linq;
+using System.Linq.Expressions;
+using Mhotivo.Models;
+
+namespace Mhotivo.App_Data.Repositories.Interfaces
+{
+    public interface IEnrollRepository
+    {
+        Enroll First(Expression<Func<Enroll, Enroll>> query);
+        Enroll GetById(long id);
+        Enroll Create(Enroll itemToCreate);
+        IQueryable<Enroll> Query(Expression<Func<Enroll, Enroll>> expression);
+        IQueryable<Enroll> Filter(Expression<Func<Enroll, bool>> expression);
+        Enroll Update(Enroll itemToUpdate, bool academicYear, bool student);
+        Enroll Delete(long id);
+        void SaveChanges();
+    }
+}
