+++ Mhotivo/App_Data/Repositories/PensumRepository.cs
@@ -2,24 +2,12 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
+using Mhotivo.App_Data.Repositories.Interfaces;
 using Mhotivo.Models;
 using System.Data.Entity;
 
 namespace Mhotivo.App_Data.Repositories
 {
-    public interface IPensumRepository : IDisposable
-    {
-        Pensum First(Expression<Func<Pensum, Pensum>> query);
-        Pensum GetById(long id);
-        Pensum Create(Pensum itemToCreate);
-        IQueryable<Pensum> Query(Expression<Func<Pensum, Pensum>> expression);
-        IQueryable<Pensum> Filter(Expression<Func<Pensum, bool>> expression);
-        Pensum Update(Pensum itemToUpdate, bool courseUpdate, bool gradeUpdate);
-        Pensum Delete(long id);
-        void SaveChanges();
-        IEnumerable<DisplayPensumModel> GetAllPesums();
-    }
-
     public class PensumRepository : IPensumRepository
     {
         private readonly MhotivoContext _context;
