+++ Mhotivo/App_Data/Repositories/StudentRepository.cs
@@ -3,27 +3,11 @@ using System.Collections.Generic;
 using System.Data.Entity;
 using System.Linq;
 using System.Linq.Expressions;
+using Mhotivo.App_Data.Repositories.Interfaces;
 using Mhotivo.Models;
 
 namespace Mhotivo.App_Data.Repositories
 {
-    public interface IStudentRepository
-    {
-        Student First(Expression<Func<Student, Student>> query);
-        Student GetById(long id);
-        Student Create(Student itemToCreate);
-        IQueryable<Student> Query(Expression<Func<Student, Student>> expression);
-        IQueryable<Student> Filter(Expression<Func<Student, bool>> expression);
-        Student Update(Student itemToUpdate);
-        Student Delete(long id);
-        Student GenerateStudentFromRegisterModel(StudentRegisterModel studentRegisterModel);
-        StudentEditModel GetStudentEditModelById(long id);
-        DisplayStudentModel GetStudentDisplayModelById(long id);
-        Student UpdateStudentFromStudentEditModel(StudentEditModel studentEditModel, Student student);
-        void SaveChanges();
-        IEnumerable<DisplayStudentModel> GetAllStudents();
-    }
-
     public class StudentRepository : IStudentRepository
     {
         private readonly MhotivoContext _context;
