+++ Mhotivo/Controllers/ClassActivityController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Web.Mvc;
 using Mhotivo.App_Data.Repositories;
+using Mhotivo.App_Data.Repositories.Interfaces;
 using Mhotivo.Logic.ViewMessage;
 using Mhotivo.Models;
