+++ Mhotivo/Controllers/EventController.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Web.Mvc;
 using System.Web.WebPages;
 using Mhotivo.App_Data.Repositories;
+using Mhotivo.App_Data.Repositories.Interfaces;
 using Mhotivo.Logic;
 using Mhotivo.Logic.ViewMessage;
 using Mhotivo.Models;
