+++ Mhotivo/Controllers/RoleController.cs
@@ -1,5 +1,6 @@
 ﻿using System.Web.Mvc;
 using Mhotivo.App_Data.Repositories;
+using Mhotivo.App_Data.Repositories.Interfaces;
 using Mhotivo.Logic.ViewMessage;
 using Mhotivo.Models;
