+++ Mhotivo/Mhotivo.csproj
@@ -150,6 +150,21 @@
     <Compile Include="App_Data\Repositories\CourseRepository.cs" />
     <Compile Include="App_Data\Repositories\EnrollRepository.cs" />
     <Compile Include="App_Data\Repositories\DiaryEventRepository.cs" />
+    <Compile Include="App_Data\Repositories\Interfaces\IAcademicYearRepository.cs" />
+    <Compile Include="App_Data\Repositories\Interfaces\IAppointmentDiaryRepository.cs" />
+    <Compile Include="App_Data\Repositories\Interfaces\IBenefactorRepository.cs" />
+    <Compile Include="App_Data\Repositories\Interfaces\IClassActivityRepository.cs" />
+    <Compile Include="App_Data\Repositories\Interfaces\IContactInformationRepository.cs" />
+    <Compile Include="App_Data\Repositories\Interfaces\ICourseRepository.cs" />
+    <Compile Include="App_Data\Repositories\Interfaces\IEnrollRepository.cs" />
+    <Compile Include="App_Data\Repositories\Interfaces\IGradeRepository.cs" />
+    <Compile Include="App_Data\Repositories\Interfaces\IMeisterRepository.cs" />
+    <Compile Include="App_Data\Repositories\Interfaces\IParentRepository.cs" />
+    <Compile Include="App_Data\Repositories\Interfaces\IPensumRepository.cs" />
+    <Compile Include="App_Data\Repositories\Interfaces\IPeopleRepository.cs" />
+    <Compile Include="App_Data\Repositories\Interfaces\IRoleRepository.cs" />
+    <Compile Include="App_Data\Repositories\Interfaces\IStudentRepository.cs" />
+    <Compile Include="App_Data\Repositories\Interfaces\IUserRepository.cs" />
     <Compile Include="App_Data\Repositories\MeisterRepository.cs" />
     <Compile Include="App_Data\Repositories\ParentRepository.cs" />
     <Compile Include="App_Data\Repositories\PensumRepository.cs" />
@@ -192,9 +207,11 @@
     <Compile Include="Logic\ViewMessage\ViewMessageLogic.cs" />
     <Compile Include="Logic\ViewMessage\ViewMessageType.cs" />
     <Compile Include="Migrations\Configuration.cs" />
+    <Compile Include="Models\AcademicYear.cs" />
     <Compile Include="Models\AcademicYearModel.cs" />
     <Compile Include="Models\AppointmentDiary.cs" />
     <Compile Include="Models\AreaModel.cs" />
+    <Compile Include="Models\Benefactor.cs" />
     <Compile Include="Models\BenefactorModel.cs" />
     <Compile Include="Models\ClassActivityGradingModel.cs" />
     <Compile Include="Models\ClassActivityModel.cs" />
@@ -202,6 +219,7 @@
     <Compile Include="Models\CourseModel.cs" />
     <Compile Include="Models\DiaryEvent.cs" />
     <Compile Include="Models\EnrollModel.cs" />
+    <Compile Include="Models\EventCreate.cs" />
     <Compile Include="Models\GradeModel.cs" />
     <Compile Include="Models\GroupModel.cs" />
     <Compile Include="Models\MeisterModel.cs" />
