+++ Mhotivo/Models/AppointmentDiary.cs
@@ -16,17 +16,4 @@ namespace Mhotivo.Models
         public User Creator { get; set; }
         public bool IsAproveed { get; set; }
     }
-
-    public class EventCreate
-    {
-        [Required(ErrorMessage = "Debe Ingresar Fecha de Inicio")]
-        [Display(Name = "Fecha de Inicio")]
-        public DateTime StartDateTime { get; set; }
-
-        [Display(Name = "Fecha de Finalización")]
-        public DateTime EndDateTime { get; set; }
-
-        [Required(ErrorMessage = "Debe Ingresar Descripción")]
-        public String Description { get; set; }
-    }
 }
