+++ Mhotivo/Models/Benefactor.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Mhotivo.Models
+{
+    public class Benefactor : People
+    {
+        public int Capacity { get; set; }
+
+        public virtual ICollection<Student> Students { get; set; }
+    }
+}
